Fix: GetMaintenanceCostFactor functions should return -1 when preconditions fail

+++ src/script/api/script_airport.hpp
@@ -198,7 +198,7 @@ public:
 	 * @pre IsAirportInformationAvailable(type)
 	 * @return Maintenance cost factor of the airport type.
 	 */
-	static uint16 GetMaintenanceCostFactor(AirportType type);
+	static int GetMaintenanceCostFactor(AirportType type);
 
 	/**
 	 * Get the monthly maintenance cost of an airport type.
